Update
This update contains the building blocks for the mesh walls of this cave. I have added a few classes that build upon each other to create the required building blocks. Please let me know if you have any ideas of better was to do this! I would love to hear. it.