Upload
Issue with the generation of the mesh. For some reason the mesh is a giant glitching square, instead of the image of the cave. When I overlay the gizmo drawing the data that the gizmo is using is correct. I have to figure out what is going wrong in the new functions I added. If anyone has any tips or feedback feel free to let me know!