Update
Debugged my program and got it to create the mesh properly. I found the following errors. The lower Control nodes where having their locations switched. so instead of it going top left, top right, bottom right, bottom left. It was going top left, top right, bottom left, bottom right. This was messing up a portion of the calculation for the mesh. The next issue that I found was that in one of my for loops, the .GetLength() method was pulling the wrong dimension to be used with the wrong iterator. This would of caused an issue when my map is rectangular.

Please let me know if anyone has an recommendations for my code!