Update
I removed the gizmo since it is no longer needed to validate that my data is correct. I also added some functions that will be responsible for determining if a side of a triangle is an edge piece. If it is determined to be an edge peace it will be marked to insure that a wall is built there.