Update
Finished the scripts for creating the wall mesh.  Found a few errors. So my next task will be to debug the code that I create and find what is causing the mesh to be created improperly. There seems to be one side of the mesh that never is getting created.

Please let me know if you have any recommendations or see where I sent wrong.